Remove state.txt concept and move indexing state into database
Getting rid of the concept of state.txt which saves the current page token of the
indexing process enables us to use multiple drives at once being mounted through one
instance of gdrivefs-rust.

Closes #9
